Fix: Correcciones de registro en base a cambios de tipo de datos en la api

+++ app/src/main/java/com/jder00138218/liftapp/network/dto/register/RegisterRequest.kt
@@ -5,7 +5,7 @@ data class RegisterRequest(
     val email: String,
     val password: String,
     val genero: String,
-    val weigth: Double,
-    val height: Double,
+    val weigth: Int,
+    val height: Int,
     val fechanac: String,
 )
